chore(types): minor adjustments to global types and general project mocks
