+++ src/components/CartList/mock.ts
@@ -4,12 +4,12 @@ const cartListMock: GameItemProps[] = [
   {
     img: '/img/halo-4.jpg',
     title: 'Halo 4',
-    price: 'R$ 200,00'
+    price: 200.0
   },
   {
     img: '/img/red-dead-img.jpg',
     title: 'Red Dead Redemption',
-    price: 'R$ 200,00'
+    price: 200.0
   }
 ]
