+++ src/components/CartList/test.tsx
@@ -8,20 +8,18 @@ import CartList from '.'
 describe('<CartList />', () => {
   it('should render the cart list', () => {
     const { container } = renderWithTheme(
-      <CartList items={cartListMock} total="R$ 400,00" />
+      <CartList items={cartListMock} total={400.0} />
     )
 
     expect(screen.getAllByRole('heading')).toHaveLength(2)
 
-    expect(screen.getByText('R$ 400,00')).toHaveStyle({ color: '#F231A5' })
+    expect(screen.getByText('$400.00')).toHaveStyle({ color: '#F231A5' })
 
     expect(container.firstChild).toMatchSnapshot()
   })
 
   it('should render the button', () => {
-    renderWithTheme(
-      <CartList items={cartListMock} total="R$ 400,00" hasButton />
-    )
+    renderWithTheme(<CartList items={cartListMock} total={400.0} hasButton />)
 
     expect(screen.getByText(/buy it now/i)).toBeInTheDocument()
   })
