+++ src/components/GameItem/index.tsx
@@ -2,6 +2,8 @@
 
 import { Download } from '@styled-icons/boxicons-solid'
 
+import formatPrice from 'utils/formatters/formatPrice'
+
 import * as S from './styles'
 
 export type PaymentInfoProps = {
@@ -14,7 +16,7 @@ export type PaymentInfoProps = {
 export type GameItemProps = {
   img: string
   title: string
-  price: string
+  price: number
   paymentInfo?: PaymentInfoProps
   downloadLink?: string
 }
@@ -46,7 +48,7 @@ const GameItem = ({
             </S.DownloadLink>
           )}
         </S.Title>
-        <S.Price>{price}</S.Price>
+        <S.Price>{formatPrice(price)}</S.Price>
       </S.Content>
     </S.GameContent>
