+++ src/dtos/home/types.ts
@@ -1,9 +1,9 @@
 import {
-  Enum_Componentpagehighlight_Alignment,
   Enum_Componentpageribbon_Color,
   Enum_Componentpageribbon_Size
 } from 'graphql/generated/graphql'
-import { Game } from '../types'
+
+import { Game, Highlight } from '../types'
 
 export type Banner = {
   title: string
@@ -14,16 +14,6 @@ export type Banner = {
   ribbonLabel?: string | null
   ribbonSize?: Enum_Componentpageribbon_Size | null
   ribbonColor?: Enum_Componentpageribbon_Color | null
-}
-
-export type Highlight = {
-  title: string
-  subtitle: string
-  buttonLabel: string
-  buttonLink: string
-  backgroundImage: string
-  floatImage?: string | null
-  alignment?: Enum_Componentpagehighlight_Alignment | null
 }
 
 export type Home = {
